Wireframe for Charity Source
Run the Charity Source .exe. There is a quirk with the first page that
will open, you won't be able to get to any other pages but you can click
on the bottom of the white page to increase the size revealing the
button to continue to the site. You can also hit the navigation button
near the top to get a list of pages. I did this all in Sketch Flow.