Users and other tweaks
Began creating the user page from the database, too sleepy to finish at
the moment.

Added jQuery and header js to the footer, its close enough to the bottom
that everything should be done rendering when we load the scripts.

Added the new table for the donations.

Random spacing fix, it was bugging me xD